+++ src/components/grid/grid-container/index.d.ts
@@ -1,30 +1,12 @@
+import propTypes from "@styled-system/space";
 import { GridItemProps } from "../grid-item";
+import { SpacingProps } from "../../../utils/helpers/options-helper";
 
-export interface GridContainerProps {
+export interface GridContainerProps extends SpacingProps {
   /** Defines the Components to be rendered within the GridContainer. Requires GridItemProps */
   children: Array<React.ReactElement<GridItemProps>> | React.ReactElement<GridItemProps>;
-  /** Any valid CSS value or a number to be multiplied by base spacing unit (8). Overrides default padding */
-  p?: string | number;
-  /** Any valid CSS value or a number to be multiplied by base spacing unit (8). Overrides default padding-left */
-  pl?: string | number;
-  /** Any valid CSS value or a number to be multiplied by base spacing unit (8). Overrides default padding-right */
-  pr?: string | number;
-  /** Any valid CSS value or a number to be multiplied by base spacing unit (8). Overrides default padding-top */
-  pt?: string | number;
-  /** Any valid CSS value or a number to be multiplied by base spacing unit (8). Overrides default padding-bottom */
-  pb?: string | number;
-  /**
-   * Any valid CSS value or a number to be multiplied by base spacing unit (8).
-   * Overrides default horizontal paddings
-   */
-  px?: string | number;
-  /**
-   * Any valid CSS value or a number to be multiplied by base spacing unit (8).
-   * Overrides default vertical paddings
-   */
-  py?: string | number;
   /** Any valid CSS value to override default grid-gap */
-  gridGap?: string;
+  gridGap?: string | number;
 }
 
 declare const GridContainer: React.FunctionComponent<GridContainerProps>;